Added SearchAction to provider, and removed method in controller.
